changed the Painter extention to Draw, updated other packages to refer it as Draw and fixed the issue with the Draw toolbar
