+++ packages/Draw/aiApps/painter/DraggableContainer.tsx
@@ -1,0 +1,62 @@
+const { useState, useRef, useEffect } = os.appHooks;
+
+const DraggableContainer = ({ children }) => {
+  const [isDragging, setIsDragging] = useState(false);
+  const [position, setPosition] = useState({
+    x: 180,
+    y: window.innerHeight / 2,
+  });
+  const [offset, setOffset] = useState({ x: 0, y: 0 });
+  const dragRef = useRef(null);
+
+  const handleMouseDown = (e) => {
+    setIsDragging(true);
+    const rect = dragRef.current.getBoundingClientRect();
+    setOffset({
+      x: e.clientX - rect.left,
+      y: e.clientY - rect.top,
+    });
+    e.preventDefault();
+  };
+
+  const handleMouseMove = (e) => {
+    if (!isDragging) return;
+    setPosition({
+      x: e.clientX - offset.x,
+      y: e.clientY - offset.y,
+    });
+  };
+
+  const handleMouseUp = () => {
+    setIsDragging(false);
+  };
+
+  useEffect(() => {
+    if (isDragging) {
+      document.addEventListener("mousemove", handleMouseMove);
+      document.addEventListener("mouseup", handleMouseUp);
+      return () => {
+        document.removeEventListener("mousemove", handleMouseMove);
+        document.removeEventListener("mouseup", handleMouseUp);
+      };
+    }
+  }, [isDragging]);
+
+  return (
+    <div
+      ref={dragRef}
+      className="draggable"
+      style={{
+        cursor: isDragging ? "grabbing" : "grab",
+        position: "fixed",
+        top: position.y,
+        left: position.x,
+      }}
+      onMouseDown={handleMouseDown}
+    >
+      {children}
+    </div>
+  );
+};
+
+export default DraggableContainer;
