+++ packages/Draw/aiApps/painter/aiApps.painter.bot.aux
@@ -4,7 +4,7 @@
       "id": "{id}",
       "space": "local",
       "tags": {
-        "forPackage": "Painter",
+        "forPackage": "Draw",
         "system": "aiApps.painter"
       }
     }
