+++ packages/Draw/aiApps/painter/closePainter.tsx
@@ -3,5 +3,5 @@ const { render } = os.appHooks;
 let prevPainter = document.getElementById("painter-container");
 
 if (prevPainter) {
-    prevPainter.remove();
+  prevPainter.remove();
 }
