+++ packages/Draw/aiApps/painter/config.json
@@ -1,0 +1,4 @@
+{
+  "author": "Sujan Barman",
+  "app": ""
+}
