+++ packages/Draw/aiApps/painter/onInstJoined.tsx
@@ -1,0 +1,9 @@
+import {
+  html2canvas,
+  ImageTracer,
+  debounce,
+} from "https://esm.helloao.org/painter-vendor-IGDNTFOW.js";
+
+globalThis.ImageTracer = ImageTracer;
+globalThis.html2canvas = html2canvas;
+globalThis.debounce = debounce;
