--- packages/Painter/aiApps/painter/config.json
@@ -1,4 +1,0 @@
-{
-    "author": "Sujan Barman",
-    "app": ""
-}
