--- packages/Painter/aiApps/painter/onInstJoined.tsx
@@ -1,9 +1,0 @@
-import {
-    html2canvas,
-    ImageTracer,
-    debounce
-} from 'https://esm.helloao.org/painter-vendor-IGDNTFOW.js';
-
-globalThis.ImageTracer = ImageTracer;
-globalThis.html2canvas = html2canvas;
-globalThis.debounce = debounce;
