+++ packages/Playlist/extension.json
@@ -1,54 +1,54 @@
 {
+  "author": "Kushagra Karki",
+  "configEditor": {
+    "app": "@tryInitPlaylistMaker",
     "author": "Kushagra Karki",
-    "configEditor": {
-      "app": "@tryInitPlaylistMaker",
-      "author": "Kushagra Karki",
-      "contextMenuConfig": {
-        "optionsIsOn": "@options"
-      },
-      "tabConfig": {
-        "title": "Playlist Kushagra"
-      },
-      "toolbarConfig": {
-        "hasToggle": "true",
-        "icon": "explore",
-        "label": "Playlist"
-      }
+    "contextMenuConfig": {
+      "optionsIsOn": "@options"
     },
-    "createdAt": "2025-09-25T16:39:37.463Z",
-    "dependencies": [
-      {
-        "depId": 1758170494705,
-        "name": "Painter",
-        "type": "package"
-      },
-      {
-        "depId": 1757954638673,
-        "name": "Scripture Map 2D",
-        "type": "package"
-      }
-    ],
-    "description": "Playlist for seed bible.",
-    "id": 1758818377463,
-    "license": "MIT",
-    "linkedDependencies": [],
-    "mainBotTag": "playlist.playlistMode",
-    "name": "Playlist",
-    "otherBots": [
-      {
-        "description": "Locale Storage",
-        "tag": "storage.tempStorageBot"
-      }
-    ],
-    "recordFile": {
-      "sha256Hash": "f1f7565cd65966ab612ec22747d0c12eaa345b4cee2a29d66bf4a58c92f6e23c",
-      "success": true,
-      "url": "https://auth-aux-aobot-prod-filesbucket-141297942820.s3.amazonaws.com/testingPublickKey/f1f7565cd65966ab612ec22747d0c12eaa345b4cee2a29d66bf4a58c92f6e23c.json"
+    "tabConfig": {
+      "title": "Playlist Kushagra"
     },
-    "source": "https://auth-aux-aobot-prod-filesbucket-141297942820.s3.amazonaws.com/testingPublickKey/f1f7565cd65966ab612ec22747d0c12eaa345b4cee2a29d66bf4a58c92f6e23c.json",
-    "status": "active",
-    "type": "package",
-    "updatedAt": "2025-09-25T16:42:15.146Z",
-    "userAuth": "1a434287-4a50-4122-a7e9-f3c02382f7a6",
-    "version": "1.0.1"
-  }
+    "toolbarConfig": {
+      "hasToggle": "true",
+      "icon": "explore",
+      "label": "Playlist"
+    }
+  },
+  "createdAt": "2025-09-25T16:39:37.463Z",
+  "dependencies": [
+    {
+      "depId": 1758170494705,
+      "name": "Draw",
+      "type": "package"
+    },
+    {
+      "depId": 1757954638673,
+      "name": "Scripture Map 2D",
+      "type": "package"
+    }
+  ],
+  "description": "Playlist for seed bible.",
+  "id": 1758818377463,
+  "license": "MIT",
+  "linkedDependencies": [],
+  "mainBotTag": "playlist.playlistMode",
+  "name": "Playlist",
+  "otherBots": [
+    {
+      "description": "Locale Storage",
+      "tag": "storage.tempStorageBot"
+    }
+  ],
+  "recordFile": {
+    "sha256Hash": "f1f7565cd65966ab612ec22747d0c12eaa345b4cee2a29d66bf4a58c92f6e23c",
+    "success": true,
+    "url": "https://auth-aux-aobot-prod-filesbucket-141297942820.s3.amazonaws.com/testingPublickKey/f1f7565cd65966ab612ec22747d0c12eaa345b4cee2a29d66bf4a58c92f6e23c.json"
+  },
+  "source": "https://auth-aux-aobot-prod-filesbucket-141297942820.s3.amazonaws.com/testingPublickKey/f1f7565cd65966ab612ec22747d0c12eaa345b4cee2a29d66bf4a58c92f6e23c.json",
+  "status": "active",
+  "type": "package",
+  "updatedAt": "2025-09-25T16:42:15.146Z",
+  "userAuth": "1a434287-4a50-4122-a7e9-f3c02382f7a6",
+  "version": "1.0.1"
+}
